app/requirements/dev.txt
python-dotenv==0.20.0
django-debug-toolbar==3.4.0